Added new tests and adjusted some functions.
- GET listings are now operated with query params in the same function.

+++ controller/users.js
@@ -34,11 +34,11 @@ usersRouter.get('/', async (_request, response) =>
   response.json(users.map(user => user.toJSON()))
 })
 
-usersRouter.delete('/:id', async (request, response) => 
+usersRouter.delete('/', async (request, response) => 
 {
   try 
   {
-    const id = request.params.id
+    const id = request.query.id;
     const user = await User.findById(id)
     const token = tokenHelper.GetToken(request)
 
@@ -52,8 +52,8 @@ usersRouter.delete('/:id', async (request, response) =>
     if (!decodedToken.id)
       return response.status(401).json({ error: 'Missing or invalid token' })
 
-    //if (user.toString() !== decodedToken.id)
-      //return response.status(401).json({ error: 'Not authorized' })
+    if (user !== decodedToken.id)
+      return response.status(401).json({ error: 'Not authorized' })
 
     const deletedUser = await user.findByIdAndRemove(id)
     response.json(deletedUser.toJSON())
